+++ ffi/json.go
@@ -4,7 +4,7 @@ import (
 	"encoding/json/v2"
 
 	"github.com/akonwi/ard/checker"
-	"github.com/akonwi/ard/vm/runtime"
+	"github.com/akonwi/ard/runtime"
 )
 
 // Encode an Ard value into a JSON string
