+++ runtime/vm.go
