+++ vm/fs_module.go
@@ -5,7 +5,7 @@ import (
 	"os"
 
 	"github.com/akonwi/ard/checker"
-	"github.com/akonwi/ard/vm/runtime"
+	"github.com/akonwi/ard/runtime"
 )
 
 // FSModule handles ard/fs module functions
