+++ vm/http_module.go
@@ -9,7 +9,7 @@ import (
 	"time"
 
 	"github.com/akonwi/ard/checker"
-	"github.com/akonwi/ard/vm/runtime"
+	"github.com/akonwi/ard/runtime"
 )
 
 // HTTPModule handles ard/http module functions
