+++ vm/result.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 
 	"github.com/akonwi/ard/checker"
-	"github.com/akonwi/ard/vm/runtime"
+	"github.com/akonwi/ard/runtime"
 )
 
 func (vm *VM) evalResultMethod(self *runtime.Object, call *checker.FunctionCall) *runtime.Object {
