+++ vm/vm.go
@@ -6,7 +6,7 @@ import (
 	"fmt"
 
 	"github.com/akonwi/ard/checker"
-	"github.com/akonwi/ard/vm/runtime"
+	"github.com/akonwi/ard/runtime"
 )
 
 type VM struct {
